+++ onsenui/esm/elements/ons-alert-dialog.spec.js
@@ -262,7 +262,6 @@ describe('OnsAlertDialogElement', () => {
       expect(div1.isEqualNode(div2)).to.be.true;
     });
   });
-});
 
   describe('autoStyling', () => {
     it('adds \'material\' modifier on Android', () => {
@@ -272,3 +271,19 @@ describe('OnsAlertDialogElement', () => {
       ons.platform.select('');
     });
   });
+
+  describe('#maskColor', () => {
+    it('sets the mask background color when set', () => {
+      dialog.maskColor = 'pink';
+      const mask = dialog.querySelector('.alert-dialog-mask');
+      expect(mask.style.backgroundColor).to.equal('pink');
+    });
+
+    it('unsets the mask background color when undefined', () => {
+      dialog.maskColor = 'pink';
+      dialog.maskColor = undefined;
+      const mask = dialog.querySelector('.alert-dialog-mask');
+      expect(mask.style.backgroundColor).to.equal('');
+    });
+  });
+});