+++ onsenui/esm/elements/ons-popover.js
@@ -499,6 +499,15 @@ export default class PopoverElement extends BaseDialogElement {
    *   [ja]バックボタンハンドラ。[/ja]
    */
 
+  /**
+   * @property maskColor
+   * @type {String}
+   * @default rgba(0, 0, 0, 0.2)
+   * @description
+   *  [en]Color of the background mask. Default is "rgba(0, 0, 0, 0.2)".[/en]
+   *  [ja]背景のマスクの色を指定します。"rgba(0, 0, 0, 0.2)"がデフォルト値です。[/ja]
+   */
+
   connectedCallback() {
     super.connectedCallback();